requirements.txt
asgiref==3.2.5
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
